nnet_layer: Add BRAM weight storage to dense layer in parallel mode
Similar to the serial mode, partition the weight array cyclically.
Only store in BRAMs when the flag is true and the required DSPs
(N_IN * N_OUT) is evenly divisibly by the reuse factor.

Oddly, does not appear to reduce fabric usage that I can determine.
Wonder what else is going on.